% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissertateUSU.R
\name{chapter_knit}
\alias{chapter_knit}
\title{Helper Function for Multipaper}
\usage{
chapter_knit(file, options = c("--csl", "apa7.csl"))
}
\arguments{
\item{file}{the name of the file (without the suffix)}

\item{options}{options passed to pandoc}
}
\description{
Helper Function for Multipaper
}
