% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissertateUSU.R
\name{dissertateUSU}
\alias{dissertateUSU}
\title{dissertateUSU format.}
\usage{
dissertateUSU(..., highlight = NULL, citation_package = "default")
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating dissertations at Utah State University. It has been approved by the graduate school.
}
\details{
Possible arguments for the YAML header are:
\itemize{
  \item \code{title} title of the dissertation
  \item \code{author} list of authors, containing \code{name} and \code{num}
  \item \code{year} year of defense
  \item \code{degree} degree earned (e.g. "Doctor of Philosphy")
  \item \code{field} the field the degree is awarded in
  \item \code{department} the department awarding the degree
  \item \code{uni} the full name of the university awarding the degree
  \item \code{location} the location of the university
  \item \code{chair} name of the defense committee chairperson
  \item \code{committee2} name of second committee member
  \item \code{committee3} name of third committee member
  \item \code{committee4} name of fourth committee member
  \item \code{committee5} name of fifth committee member (if left blank, committee4 will be ignored and the cover page will be formatted as a masters thesis)
  \item \code{gradschool} name of the graduate school member who needs to sign off on the degree
  \item \code{gradschoollabel} the position of the graduate school member
  \item \code{bibliography} BibTeX \code{.bib} file name
  \item \code{multipaper} whether the design is a multi-paper format (true) or monograph (false). This provides additional functionality for having citations at the end of each chapter.
  \item \code{csl} the .csl citation style file name
  \item \code{header-includes}: custom additions to the header, before the \code{\\begin\{document\}} statement
  \item \code{include-after}: for including additional LaTeX code before the \code{\\end\{document\}} statement
  }
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "dissertateUSU", package = "dissertateUSU")
}

}
