% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preamble_for_title.R
\name{preamble_for_title}
\alias{preamble_for_title}
\title{Preamble for Title Page}
\usage{
preamble_for_title(file = "preamble.tex")
}
\description{
Creates the preamble.tex file for the title page to work
}
